python 3.10 updates. Deprecate np.object.

+++ chroma/bvh/bvh.py
@@ -6,7 +6,7 @@ from pycuda.gpuarray import vec
 uint4 = vec.uint4 # pylint: disable-msg=C0103,E1101
 
 CHILD_BITS = 28
-NCHILD_MASK = np.uint32(0xFFFF << CHILD_BITS)
+NCHILD_MASK = np.uint64(0xFFFF << CHILD_BITS)
 
 def unpack_nodes(nodes):
     '''Creates a numpy record array with the contents of nodes
